Create mylarpubtoconfig.py
This script reads from the defined mylar.db in config (or configPRIVATE) to pull publishers that exist in Mylar's DB of added comics. Then it places all those publisher IDs into the config for processing with yearinreview.py 

Intent is to allow users to see what series they are missing from their already preferred publishers in a more automated fashion.